2023/21: Diamond Maze
This was the crappiest AoC challenge ever. Not only does the example
look completely different than the input data, it also can't be
optimized in the same way. You have to look at the input data to figure
out how to even solve this. And you have to know how polynomials work or
end up with a horribly messy solution. The worst challenge I've done
here, complete disaster...